Preliminary Dataset and Model Screening/breast_mri_training/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=8.0.0
tqdm>=4.60.0
scikit-learn>=1.0.0
matplotlib>=3.3.0
nibabel>=3.2.0
pydicom>=2.3.0
pynrrd>=1.0.0
pandas>=1.3.0
pyarrow>=6.0.0

